+++ test/functional/tools/system/package_manager_test.py
@@ -15,7 +15,7 @@ def test_apt_check():
     client = TestClient()
     client.save({"conanfile.py": textwrap.dedent("""
         from conan import ConanFile
-        from conan.tools.system.package_manager import Apt
+        from conan.tools.system import Apt
 
         class MyPkg(ConanFile):
             settings = "arch", "os"
@@ -34,7 +34,7 @@ def test_apt_install_substitutes():
     client = TestClient()
     conanfile_py = textwrap.dedent("""
         from conan import ConanFile
-        from conan.tools.system.package_manager import Apt
+        from conan.tools.system import Apt
         class MyPkg(ConanFile):
             settings = "arch", "os"
             def system_requirements(self):
@@ -67,7 +67,7 @@ def test_build_require():
     client = TestClient()
     client.save({"tool_require.py": textwrap.dedent("""
         from conan import ConanFile
-        from conan.tools.system.package_manager import Apt
+        from conan.tools.system import Apt
 
         class MyPkg(ConanFile):
             settings = "arch", "os"
@@ -94,7 +94,7 @@ def test_brew_check():
     client = TestClient()
     client.save({"conanfile.py": textwrap.dedent("""
         from conan import ConanFile
-        from conan.tools.system.package_manager import Brew
+        from conan.tools.system import Brew
 
         class MyPkg(ConanFile):
             settings = "arch"
@@ -114,7 +114,7 @@ def test_brew_install_check_mode():
     client = TestClient()
     client.save({"conanfile.py": textwrap.dedent("""
         from conan import ConanFile
-        from conan.tools.system.package_manager import Brew
+        from conan.tools.system import Brew
 
         class MyPkg(ConanFile):
             settings = "arch"
@@ -134,7 +134,7 @@ def test_brew_install_install_mode():
     client = TestClient()
     client.save({"conanfile.py": textwrap.dedent("""
         from conan import ConanFile
-        from conan.tools.system.package_manager import Brew
+        from conan.tools.system import Brew
 
         class MyPkg(ConanFile):
             settings = "arch"
@@ -154,7 +154,7 @@ def test_collect_system_requirements():
     client = TestClient()
     client.save({"conanfile.py": textwrap.dedent("""
         from conan import ConanFile
-        from conan.tools.system.package_manager import Brew, Apt
+        from conan.tools.system import Brew, Apt
 
         class MyPkg(ConanFile):
             settings = "arch"
