+++ test/integration/tools/system/package_manager_test.py
@@ -5,8 +5,8 @@ import mock
 import pytest
 from unittest.mock import PropertyMock
 
-from conan.tools.system.package_manager import Apt, Apk, Dnf, Yum, Brew, Pkg, PkgUtil, Chocolatey, \
-    Zypper, PacMan, _SystemPackageManagerTool
+from conan.tools.system import Apt, Apk, Dnf, Yum, Brew, Pkg, PkgUtil, Chocolatey, Zypper, PacMan
+from conan.tools.system.package_manager import _SystemPackageManagerTool
 from conan.errors import ConanException
 from conan.internal.model.settings import Settings
 from conan.test.utils.mocks import ConanFileMock, MockSettings
